+++ frontends/web/src/containers/CreateInterface.js
@@ -34,6 +34,9 @@ import {
   OverlayContext,
   BadgeOverlay
 } from "./Overlay"
+import {
+  HateSpeechDropdown
+} from "./HateSpeechDropdown.js"
 import "./CreateInterface.css";
 
 const Explainer = (props) => (
@@ -102,17 +105,6 @@ const GoalMessage = ({ targets = [], curTarget, taskType, taskShortName, onChang
         }
       </div>
     </div>
-  );
-};
-
-const HateSpeechDropdown = ({ hateType, dataIndex, onClick }) => {
-  return (
-    <DropdownButton variant="light" size="sm" className="p-1" title={hateType ? "Type of hate: " + hateType : "Type of hate"}>
-      {
-        ["Threatening language", "Supporting hateful entities", "Derogation", "Dehumanizing language", "Animosity", "None selected"].map((type, index) =>
-        <Dropdown.Item data-index={dataIndex} data={type} onClick={onClick} key={index} index={index}>{type}</Dropdown.Item>)
-      }
-    </DropdownButton>
   );
 };
